Moved model classes into model package
