+++ src/server/model/ChatServer.java
@@ -1,4 +1,4 @@
-package server;
+package server.model;
 
 import java.io.BufferedReader;
 import java.io.IOException;
