+++ src/server/model/Message.java
@@ -1,4 +1,4 @@
-package server;
+package server.model;
 
 public class Message {
