+++ src/server/model/MessageSender.java
@@ -1,4 +1,4 @@
-package server;
+package server.model;
 
 import java.io.IOException;
 import java.io.PrintWriter;
