+++ src/server/model/ServerConnection.java
@@ -1,4 +1,4 @@
-package server;
+package server.model;
 
 import java.io.IOException;
 import java.net.ServerSocket;
